+++ benches/src/import.rs
@@ -1,0 +1,7 @@
+mod pressure_block_importer_port;
+mod pressure_consensus_port;
+mod pressure_peer_to_peer_port;
+
+pub use pressure_block_importer_port::PressureBlockImporterPort;
+pub use pressure_consensus_port::PressureConsensusPort;
+pub use pressure_peer_to_peer_port::PressurePeerToPeerPort;
